Add files via upload
 CSV file has three columns: transmission time in milliseconds, signal amplitude, and
FM radio frequency in kHz. One must find out which radio frequency that
is generating a structured signal and which is just random noise. Once
we’ve found the correct frequency, then I do a scatter plot of the data that
was transmitted. Finally, I’ll need to draw a line plot of the mean
amplitude for each unit time on top of the scatter plot.